refactor: use client-safe sortCommunitiesCustom import to resolve Next.js dev/build webpack issues
